\name{migration.gini.in.standardized}
\alias{migration.gini.in.standardized}
\title{.. content for}
\usage{
  migration.gini.in.standardized(m)
}
\arguments{
  \item{m}{}
}
\description{
  .. content for
}
\author{
  Gergely Daróczi
}
\references{
  Rogers, A. and Raymer, J. (1998) The spatial focus of US
  interstate migration flows. Int. J PoplnGeogr,4, 63-80.
}

