\name{migration.gini}
\alias{migration.gini}
\title{Spatial Gini Indexes}
\usage{
  migration.gini(m)
}
\arguments{
  \item{m}{migration matrix}
}
\description{
  Spatial Gini Indexes
}
\examples{
\dontrun{
data(migration.hyp)
migration.gini(migration.hyp)
migration.gini(migration.hyp2)
}
}
\seealso{
  \code{\link{migration.gini.col}}
  \code{\link{migration.gini.row}}
  \code{\link{migration.gini.exchange}}
  \code{\link{migration.gini.in}}
  \code{\link{migration.gini.out}}
}

