\docType{data}
\name{migration.hyp}
\alias{migration.hyp}
\title{Hypotetical Migration Matrix}
\format{migration matrix}
\description{
  A small (3x3) hypotetical migration matrix.
}
\references{
  David A. Plane and Gordon F. Mulligan (1997): Measuring
  Spatial Focusing in a Migration System. In. Demography,
  Vol. 34, No. 2 (May, 1997), pp. 253
}
\keyword{data}
\keyword{datasets}

