% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.indices.R
\docType{data}
\name{migration.world}
\alias{migration.world}
\title{Global Bilateral Migration Database (2000)}
\format{
migration matrix
}
\description{
Global (country-to-country) matrix of bilateral migrant stocks in 2000 with 226 economies involved.
}
\references{
\itemize{
  \item World Bank (2010): Global Bilateral Migration Database.
}
}
\keyword{data}
