\name{migration.gini}
\alias{migration.gini}
\title{Spatial Gini Indexes}
\usage{
  migration.gini(m)
}
\arguments{
  \item{m}{migration matrix}
}
\description{
  Spatial Gini Indexes
}
\examples{
\dontrun{
m <- matrix(c(0, 20, 30, 10, 0, 30, 10, 20, 0),3,3)
migration.gini(m)
}
}
\author{
  Gergely Daróczi
}

