\docType{data}
\name{migration.hyp}
\alias{migration.hyp}
\alias{migration.hyp2}
\title{Hypotetical Migration Matrix}
\format{migration matrix}
\description{
  A small (3x3) hypotetical migration matrix.
}
\references{
  \itemize{ \item David A. Plane and Gordon F. Mulligan
  (1997): Measuring Spatial Focusing in a Migration System.
  \emph{Demography} \bold{34}, pp. 253 \item Andrei Rogers
  and Stuart Sweeney (1998) Measuring the Spatial Focus of
  Migration Patterns. \emph{The Professional Geographer}
  \bold{50}, 232--242 }
}
\keyword{data}
\keyword{datasets}

