\docType{package}
\name{migration.indices}
\alias{migration.indices}
\alias{migration.indices-package}
\title{\emph{Migration indices}}
\description{
  This package provides various indices, like Crude
  Migration Rate, different Gini indices or the Coefficient
  of Variation among others, to show the (un)equality of
  migration.
}

