\docType{data}
\name{migration.world}
\alias{migration.world}
\title{Global Bilateral Migration Database (2000)}
\format{migration matrix}
\description{
  Global (country-to-country) matrix of bilateral migrant
  stocks in 2000 with 226 economies involved.
}
\references{
  \itemize{ \item World Bank (2010): Global Bilateral
  Migration Database.
  \url{http://data.worldbank.org/data-catalog/global-bilateral-migration-database}
  }
}
\keyword{data}
\keyword{datasets}

